+++ bin/main.dart
@@ -7,11 +7,7 @@ import 'package:trcli/translate_cli.dart';
 
 import '../lib/src/io/commands.dart';
 
-final _parser = ArgParser();
-
 Future<void> main(List<String> args) async {
-  // testMapper();
-  // return;
   var runner = CommandRunner(
     'trcli',
     'cli to make your app\'s l10n easy',
@@ -27,37 +23,6 @@ Future<void> main(List<String> args) async {
   }
 }
 
-// var keys = {};
-
-// void testMapper() {
-//   // var str = 'Welcome back {{user}}, today is {{date}}.';
-//   // var params = captureVars(str);
-//   // str = params.text;
-//   // trace('super intput ' ,str);
-//   // var output = replaceVars(params);
-//   // trace('super output ' ,output);
-//   // trace(params.text);
-//   // trace(params.vars);
-// }
-
-// String trArgs2(List<dynamic> params) {
-//   var str = tr;
-//   if (_matchParamsRegExp.hasMatch(str)) {
-//     // deduplicate matches
-//     final wordset = <String>{};
-//     final matches = _matchParamsRegExp.allMatches(str);
-//     for (var match in matches) {
-//       wordset.add(str.substring(match.start, match.end));
-//     }
-//     // Replacing
-//     var words = wordset.toList();
-//     for (var i = 0; i < words.length; i++) {
-//       str = str.replaceAll(words[i], '${params[i]}');
-//     }
-//   }
-//   return str;
-// }
-
 void startRun() {
   // trace("Start run!");
   build();
@@ -70,6 +35,8 @@ Future<void> startFetch() async {
   trace('Fetching data from Google sheets...');
   final localesMap = await sheet.getData();
   localesMap[config.masterLocale] = canoMap;
+  putVarsInMap(localesMap);
+
   if (config.validTKeyFile) {
     createTKeyFileFromMap(map, save: true, includeToString: false);
   }